Refactor field components structure
