+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/component/field/POFieldLabels.kt
@@ -1,6 +1,7 @@
 package com.processout.sdk.ui.core.component.field
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Immutable
 import com.processout.sdk.ui.core.annotation.ProcessOutInternalApi
 import com.processout.sdk.ui.core.component.POText
 import com.processout.sdk.ui.core.style.input.POInputStateStyle
@@ -9,12 +10,24 @@ import com.processout.sdk.ui.core.theme.ProcessOutTheme
 
 /** @suppress */
 @ProcessOutInternalApi
-object POFieldLabelsDefaults {
+object POFieldLabels {
 
-    val default: POFieldLabelsStyle
+    @Immutable
+    data class Style(
+        val normal: StateStyle,
+        val error: StateStyle
+    )
+
+    @Immutable
+    data class StateStyle(
+        val title: POText.Style,
+        val description: POText.Style
+    )
+
+    val default: Style
         @Composable get() = with(ProcessOutTheme) {
-            POFieldLabelsStyle(
-                normal = POFieldLabelsStateStyle(
+            Style(
+                normal = StateStyle(
                     title = POText.Style(
                         color = colors.text.secondary,
                         textStyle = typography.fixed.labelHeading
@@ -24,7 +37,7 @@ object POFieldLabelsDefaults {
                         textStyle = typography.fixed.label
                     )
                 ),
-                error = POFieldLabelsStateStyle(
+                error = StateStyle(
                     title = POText.Style(
                         color = colors.text.secondary,
                         textStyle = typography.fixed.labelHeading
@@ -38,13 +51,13 @@ object POFieldLabelsDefaults {
         }
 
     @Composable
-    fun custom(style: POInputStyle) = POFieldLabelsStyle(
+    fun custom(style: POInputStyle) = Style(
         normal = style.normal.toStateStyle(),
         error = style.error.toStateStyle()
     )
 
     @Composable
-    private fun POInputStateStyle.toStateStyle() = POFieldLabelsStateStyle(
+    private fun POInputStateStyle.toStateStyle() = StateStyle(
         title = POText.custom(style = title),
         description = POText.custom(style = description)
     )
